requirements.txt
# Core Framework
langgraph==0.2.45
langchain==0.3.7
langchain-openai==0.2.9
langchain-core==0.3.15

# API Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic>=2.9.0
pydantic-settings>=2.6.0

# OpenAI
openai==1.51.0

# Video Processing
moviepy==1.0.3
Pillow==10.4.0

# Audio Processing
pydub==0.25.1

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
aiofiles==24.1.0

# Database (SQLite with SQLAlchemy)
sqlalchemy==2.0.35
alembic==1.13.3
